Move components into standalone
