+++ source/renderer/App.tsx
@@ -15,12 +15,12 @@ import { PasswordPrompt } from "./components/PasswordPrompt";
 import { AddVaultMenu } from "./components/AddVaultMenu";
 import { PreferencesDialog } from "./components/PreferencesDialog";
 import { Notifications } from "./components/Notifications";
-import { FileHostConnectionNotice } from "./components/FileHostConnectionNotice";
-import { CreateNewFilePrompt } from "./components/CreateNewFilePrompt";
+import { FileHostConnectionNotice } from "./components/standalone/FileHostConnectionNotice";
+import { CreateNewFilePrompt } from "./components/standalone/CreateNewFilePrompt";
 import { UpdateDialog } from "./components/standalone/UpdateDialog";
 import { AboutDialog } from "./components/standalone/AboutDialog";
 import { VaultManagementDialog } from "./components/standalone/VaultManagementDialog";
-import { BiometricRegistrationDialog } from "./components/BiometricRegistrationDialog";
+import { BiometricRegistrationDialog } from "./components/standalone/BiometricRegistrationDialog";
 import { useTheme } from "./hooks/theme";
 import { Theme } from "./types";
 import { userCopiedText } from "./actions/clipboard";
