+++ source/renderer/components/standalone/FileHostConnectionNotice.tsx
@@ -2,10 +2,10 @@ import * as React from "react";
 import { useState as useHookState } from "@hookstate/core";
 import styled from "styled-components";
 import { Button, Card, Classes, Dialog } from "@blueprintjs/core";
-import { FILE_HOST_CODE } from "../state/fileHost";
-import { copyText } from "../actions/clipboard";
-import { showSuccess } from "../services/notifications";
-import { t } from "../../shared/i18n/trans";
+import { FILE_HOST_CODE } from "../../state/fileHost";
+import { copyText } from "../../actions/clipboard";
+import { showSuccess } from "../../services/notifications";
+import { t } from "../../../shared/i18n/trans";
 
 const { useCallback } = React;
